elf: Use the get_file_size function
Simplified retrieval of file size by using the function get_file_size.

+++ src/include/rimage/rimage.h
@@ -61,7 +61,7 @@ struct module {
 	int data_file_size;
 
 	/* total file size */
-	int file_size;
+	size_t file_size;
 
 	/* executable header module */
 	int exec_header;
